// frontend/src/app/shared/models/usuario-model.ts
export interface Usuario {
  _id?: any;
  cpf?: string;
  usuario?: string;
  rg?: string;
  cns?: string;
  sexo?: string;
  data_nasc?: any;
  n_celular?: string;
  idade?: any;
  cep?: string;
  logradouro?: string;
  numero?: number;
  bairro?: string;
  uf?: string;
  cidade?: string;
  termos?: boolean;
  validado?: boolean;
  descartado?: boolean;
  atendido?: boolean;
  nome?: string;

  ubs?: string;
  atividade?: string;
  procedimento?: string;
  detalhe?: string;
  data?: string;
  dia?: string;
  horas?: string;
  hr?: any;
  prioridade?: string;
  categoria?: string;

}